Explain how users can contribute to the documentation
It took me a while to figure out how to do this. I only stumbled across it by going to 'github.com/fish-shell' and seeing 'The website for fishshell.com' as a separate repo.
Not sure which section you might want this in ... I've used this page 'guide for developers' but I guess non-devs can also edit the docs.