Test the emoji width at compiletime
Since Unicode 9, the width of some characters changed to 2.

Depending on the system, it might have support for it, or it might
not.

Instead of hardcoding specific glibc etc versions, we compile a small
program and check an actual value.

The intention is to, in most cases, make setting $fish_emoji_width
unnecessary.

Unfortunately this approach has some caveats:

- It relies on the locale being set to a unicode-supporting one.
  (C.UTF-8 is unfortunately not standard, so we can't use it)
- It relies on the terminal's wcwidth having unicode9 support IFF the
  system wcwidth does.

Some results:

Ubuntu Xenial (16.04) on Travis correctly detects 1 (glibc is 2.23,
the cut-off for unicode9 support is 2.26).

macOS should mostly be unaffected as we have special heuristics for
iTerm and Terminal.app. Regardless, the result is 2 for macOS 10.13.3.

Alpine edge (which uses musl) also returns 2.