style guide: allow multi-line comments
I think our style guide is overly prescriptive. It says all comments must be "C++ style"; as in starting with `//` (or `///` for documenting code). Multi-line/block comments in the style of `/* ... */` are disallowed. C++ is fine with these kinds of comments, and they are nicer to use when a comment is multiple lines of text. So let's not do that.