Revert "Optimize function calls by reducing inherit vars heap allocations and copies"
This reverts commit cdce8511a14e0a84f3796cce6889c3ce005d40b1.

This change was unsafe. The prior version (now restored) took the lock and
then copied the data. By returning a reference, the caller holds a
reference to data outside of the lock.

This function isn't worth optimizing. Hardly any functions use this
facility, and for those that do, they typically just capture one or two
variables.