Reformat all(*) fish function files
This just runs

```fish
for f in **.fish
    fish_indent -w $f
end
```

on the entire tree. We have a formatting tool, we should use it. If it
makes some things look weird, we should fix it or live with it looking
slightly weird.

(*): The exceptions are:

- fish_vi_cursor, because that has a PR that I'd like to merge first

- completions/magento, because that looks weird in `git diff`, and I
want to look at it some more.